Actualizando el repo

+++ src/Gestiona360.Payroll.Application.Contracts/DTOs/Employees/EmployeeGarnishmentLineDto.cs
@@ -11,7 +11,7 @@ namespace Gestiona360.Payroll.Application.Contracts.DTOs.Employees
         public Guid Id { get; set; }
         public string Type { get; set; } = string.Empty; // Alimony, Civil
         public string CourtOrderNumber { get; set; } = string.Empty;
-        public int PercentageLimit { get; set; }
+        public decimal PercentageLimit { get; set; }
         public decimal TotalAmountToWithhold { get; set; }
         public decimal WithheldToDate { get; set; }
         public bool IsActive { get; set; }
